+++ Box.h
@@ -5,17 +5,16 @@
 #ifndef PROJECT101_BOX_H
 #define PROJECT101_BOX_H
 
-
 #include <array>
 
 class Box {
 public:
     Box();
-    Box(std::array<int, 2> c, std::array<int, 2> h);
+    Box(std::array<double, 2> c, std::array<double, 2> h);
     bool Overlaps(Box other);
 
-    std::array<int, 2> center;
-    std::array<int, 2> halfSize;
+    std::array<double, 2> center;
+    std::array<double, 2> halfSize;
 };
